+++ srcs/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: gpacheco <gpacheco@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/22 10:50:07 by gpacheco          #+#    #+#             */
-/*   Updated: 2022/03/23 09:54:33 by gpacheco         ###   ########.fr       */
+/*   Updated: 2022/03/29 17:28:43 by gpacheco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int		print_parameter(char *pointer, va_list args);
 
 char	*ft_itoa(int n);
 char	*ft_utoa(unsigned int n);
-char	*ft_itoh(intptr_t n, int cap);
+char	*ft_itoh(unsigned long int n, int cap);
 int		ft_isprint(int c);
 size_t	ft_strlen(const char *c);
